requirements.txt
requests
schedule
python-dotenv
yfinance
